Command Palette
This is inspired by Sublime Text. The command palette can give quick access to
every command in the application. It was introduced to make it easier to add new
commands (which may take parameters), because adding proper UI (menus, dialogs
etc.) is complex.

This way, new or experimental features may first be added in the command
palette, and can later be integrated in the UI proper.