Steps towards a FeatureIDE client API
The idea is to run operations (such as adding a feature) not only on the server,
but on the client as well to enable optimistic UI. It is desirable to use the
same operation code on client and server to ensure consistent behaviour. JSweet
transpiles the Java operations to JavaScript.

As the operations use FeatureIDE APIs (IFeatureModel and friends), a part of
FeatureIDE has to be re-implemented on the client (see README). For the core
API (managing feature trees) this is not actually that hard to do, see
bridge.ts.